+++ R/data_battles.R
@@ -11,8 +11,4 @@
 #'   \item{US_party}{is the US a party to the battle, numeric}
 #'   \item{N_actors}{number of actors to conflict, numeric}
 #' }
-#' @details
-#' ``` {r, echo = FALSE, warning = FALSE}
-#' mreport(battles)
-#' ```
 "battles"