+++ R/operate_proportional.R
@@ -1,17 +1,17 @@
 #' Proportion of messy dates meeting logical test
-#'
-#' These functions provide various proportional tests for messy date objects.
-#' @name proportional
+#' @description
+#'   These functions provide various proportional tests for messy date objects.
+#' @name operate_proportional
 #' @param e1,e2 `mdate` or other class objects
 #' @return The proportion that the comparison is true.
 #' @return A logical vector the same length as the `mdate` passed.
 NULL
 
-#' @rdname proportional
+#' @rdname operate_proportional
 #' @export
 `%l%` <- function(e1, e2) UseMethod("%l%")
 
-#' @describeIn proportional Tests proportion of dates in the first vector
+#' @describeIn operate_proportional Tests proportion of dates in the first vector
 #'   that precede the minimum in the second vector.
 #' @examples
 #'   as_messydate("2012-06") < as.Date("2012-06-02")
@@ -30,11 +30,11 @@ evalqOnLoad({
   registerS3method("%l%", "POSIXt", `%l%.mdate`)
 })
 
-#' @rdname proportional
+#' @rdname operate_proportional
 #' @export
 `%g%` <- function(e1, e2) UseMethod("%g%")
 
-#' @describeIn proportional Tests proportion of dates in the first vector
+#' @describeIn operate_proportional Tests proportion of dates in the first vector
 #'   that follow the maximum in the second vector.
 #' @export
 #' @examples
@@ -53,11 +53,11 @@ evalqOnLoad({
   registerS3method("%g%", "POSIXt", `%g%.mdate`)
 })
 
-#' @rdname proportional
+#' @rdname operate_proportional
 #' @export
 `%ge%` <- function(e1, e2) UseMethod("%ge%")
 
-#' @describeIn proportional Tests proportion of dates in the first vector
+#' @describeIn operate_proportional Tests proportion of dates in the first vector
 #'   that follow or are equal to the maximum in the second vector.
 #' @export
 #' @examples
@@ -76,11 +76,11 @@ evalqOnLoad({
   registerS3method("%ge%", "POSIXt", `%ge%.mdate`)
 })
 
-#' @rdname proportional
+#' @rdname operate_proportional
 #' @export
 `%le%` <- function(e1, e2) UseMethod("%le%")
 
-#' @describeIn proportional Tests proportion of dates in the first vector
+#' @describeIn operate_proportional Tests proportion of dates in the first vector
 #'   that precede or are equal to the minimum in the second vector.
 #' @export
 #' @examples
@@ -99,11 +99,11 @@ evalqOnLoad({
   registerS3method("%le%", "POSIXt", `%le%.mdate`)
 })
 
-#' @rdname proportional
+#' @rdname operate_proportional
 #' @export
 `%><%` <- function(e1, e2) UseMethod("%><%")
 
-#' @describeIn proportional Tests proportion of dates in the first vector
+#' @describeIn operate_proportional Tests proportion of dates in the first vector
 #'   that are between the minimum and maximum dates in the second vector.
 #' @export
 #' @examples
@@ -123,11 +123,11 @@ evalqOnLoad({
   registerS3method("%><%", "POSIXt", `%><%.mdate`)
 })
 
-#' @rdname proportional
+#' @rdname operate_proportional
 #' @export
 `%>=<%` <- function(e1, e2) UseMethod("%>=<%")
 
-#' @describeIn proportional Tests proportion of dates in the first vector that
+#' @describeIn operate_proportional Tests proportion of dates in the first vector that
 #'   are between the minimum and maximum dates in the second vector, inclusive.
 #' @export
 #' @examples