+++ tests/testthat/test-coerce_to.R
@@ -1,7 +1,7 @@
 test_that("Coercion from other date classes into messydt works", {
   date <- as.Date("2010-10-10")
-  POSIXct <- as.POSIXct("2010-10-10")
-  POSIXlt <- as.POSIXlt("2010-10-10")
+  POSIXct <- as.POSIXct("2010-10-10", tz = "UTC")
+  POSIXlt <- as.POSIXlt("2010-10-10", tz = "UTC")
   character <- "2010-10-10"
   character2 <- "AD2010-10-10"
   character3 <- "{BC2010-10-10,BC2010-10-11,BC2010-10-12}"